src/requirements.txt
Flask
psycopg2-binary
gunicorn
python-dotenv
pandas
setuptools
wheel
flasgger